requirements.txt
requests>=2.28.0
pyyaml>=6.0
pydantic>=2.0.0
jinja2>=3.1.0
boto3>=1.28.0
